[FIX] mail_show_follower: Company identification in _send()
Before this commit, 'Show Internal Users CC' would always be False,
regardless of the setting in the company, in case the message came
from a record with a company_id field but without a value assignment to
it (e.g., a partner record with no company), as company would be an
empty record.

This commit fixes the behavior by capturing company from self.env in
these cases.